DataViews: Use ControlWithError from @wordpress/ui
Replaces the vendored ControlWithError and ValidityIndicator copies with
the @wordpress/ui port, as planned when they were vendored. The
validated controls keep the dataviews-validated-control wrapper class
(now passed as a className) so the :has()-based invalid styles keep
applying; the indicator styles now ship with the ui component, so their
SCSS copy is removed. The pending-validation test moves into the
ValidatedSelectControl test file.

Note: this branch is based on add/wp-ui-spinner, which predates the
vendored ValidatedToggleControl (#81492). Once trunk is merged back into
this branch, packages/dataviews/src/components/validated-form-controls/
toggle-control.tsx must also be switched to the @wordpress/ui import.

Part of https://github.com/WordPress/gutenberg/issues/81230
