+++ packages/dataviews/src/components/validated-form-controls/form-token-field.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, useRef } from '@wordpress/element';
 import { FormTokenField } from '@wordpress/components';
-import { ControlWithError } from './control-with-error';
+import { ControlWithError } from '@wordpress/ui';
 import type { ValidatedControlProps } from './types';
 
 type FormTokenFieldProps = React.ComponentProps< typeof FormTokenField >;
@@ -24,6 +24,7 @@ const UnforwardedValidatedFormTokenField = (
 			ref={ forwardedRef }
 		>
 			<ControlWithError
+				className="dataviews-validated-control"
 				required={ required }
 				markWhenOptional={ markWhenOptional }
 				customValidity={ customValidity }
