+++ packages/dataviews/src/components/validated-form-controls/number-control.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, useRef } from '@wordpress/element';
 import { useMergeRefs } from '@wordpress/compose';
 import { __experimentalNumberControl as NumberControl } from '@wordpress/components';
-import { ControlWithError } from './control-with-error';
+import { ControlWithError } from '@wordpress/ui';
 import type { ValidatedControlProps } from './types';
 
 type NumberControlProps = React.ComponentProps< typeof NumberControl >;
@@ -22,6 +22,7 @@ const UnforwardedValidatedNumberControl = (
 
 	return (
 		<ControlWithError
+			className="dataviews-validated-control"
 			required={ required }
 			markWhenOptional={ markWhenOptional }
 			customValidity={ customValidity }
