+++ packages/dataviews/src/components/validated-form-controls/style.scss
@@ -49,45 +49,3 @@
 	pointer-events: none;
 }
 
-.dataviews-validated-control__indicator {
-	display: flex;
-	align-items: flex-start;
-	gap: 4px;
-	margin: 8px 0 0;
-	font-family: $font-family-body;
-	font-size: $font-size-small;
-	line-height: 16px; // matches the icon size
-	color: $gray-700;
-	animation:
-		dataviews-validated-control__indicator-jump 0.2s
-		cubic-bezier(0.68, -0.55, 0.27, 1.55);
-
-	&.is-invalid {
-		color: $alert-red;
-	}
-
-	&.is-valid {
-		color: color-mix(in srgb, $black 30%, $alert-green);
-	}
-}
-
-.dataviews-validated-control__indicator-icon {
-	flex-shrink: 0;
-}
-
-.dataviews-validated-control__indicator-spinner {
-	margin: 2px;
-	width: $grid-unit-15;
-	height: $grid-unit-15;
-}
-
-@keyframes dataviews-validated-control__indicator-jump {
-	0% {
-		transform: translateY(-4px);
-		opacity: 0;
-	}
-	100% {
-		transform: translateY(0);
-		opacity: 1;
-	}
-}
