+++ packages/dataviews/src/components/validated-form-controls/test/select-control.tsx
@@ -1,5 +1,6 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { useRef } from '@wordpress/element';
 import { ValidatedSelectControl } from '../select-control';
 
 describe( 'ValidatedSelectControl', () => {
@@ -53,5 +54,57 @@ describe( 'ValidatedSelectControl', () => {
 		expect( select ).toHaveAccessibleDescription(
 			expect.stringContaining( 'Pick a color.' )
 		);
+	} );
+
+	it( 'should keep the pending indicator instead of a native error on a synthetic `invalid` event', async () => {
+		const user = userEvent.setup();
+
+		function PendingValidatedSelectControl() {
+			const ref = useRef< HTMLSelectElement >( null );
+			return (
+				<>
+					<ValidatedSelectControl
+						ref={ ref }
+						label="Color"
+						required
+						value=""
+						options={ [
+							{ label: 'Select a color...', value: '' },
+							{ label: 'Red', value: 'red' },
+						] }
+						onChange={ () => {} }
+						customValidity={ {
+							type: 'validating',
+							message: 'Validating...',
+						} }
+					/>
+					<button
+						type="button"
+						onClick={ () =>
+							ref.current?.dispatchEvent(
+								new Event( 'invalid', {
+									cancelable: true,
+								} )
+							)
+						}
+					>
+						Show errors
+					</button>
+				</>
+			);
+		}
+
+		render( <PendingValidatedSelectControl /> );
+
+		await user.click(
+			screen.getByRole( 'button', { name: 'Show errors' } )
+		);
+
+		await waitFor( () => {
+			expect( screen.getByText( 'Validating...' ) ).toBeVisible();
+		} );
+		expect(
+			screen.queryByText( 'Constraints not satisfied' )
+		).not.toBeInTheDocument();
 	} );
 } );
